% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImport.R
\name{dataImport_demo}
\alias{dataImport_demo}
\title{Data import demo}
\usage{
dataImport_demo()
}
\value{
A Shiny App.
}
\description{
R Shiny module demo to import data files
}
\details{
This module allow to import multiple type of data.
The file type currently supported are csv, txt, xls, xslx, rda.
This demo allow the user to select a dataframe from a file and render it.
}
\examples{
\dontrun{
 dataImport_demo()
}
}
\keyword{dataframe}
