% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImport.R
\name{dataImport_server}
\alias{dataImport_server}
\title{Data import server}
\usage{
dataImport_server(id)
}
\arguments{
\item{id}{A string.}
}
\value{
A Shiny server.
}
\description{
R Shiny module server to import data files
}
\details{
This module allow to import multiple type of data.
The file type currently supported are csv, txt, xls, xslx, rda.
The server dynamically create a selection input if multiple
dataframe are present in the file selected.
}
\examples{
\dontrun{
    dataImport_demo()
}
}
\keyword{data}
