% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImport.R
\name{dataImport_ui}
\alias{dataImport_ui}
\title{Data import ui}
\usage{
dataImport_ui(id, label = "Select data file")
}
\arguments{
\item{id}{A string.}

\item{label}{A string use to prompt the user}
}
\value{
A Shiny UI.
}
\description{
R Shiny module UI to import data files
}
\details{
This module allow to import multiple type of data.
The file type currently supported are csv, txt, xls, xslx, rda.
The UI ask the user for the file localisation, the separator,
the needs to format to character, the quote format, the presence
of heading, the conversion of string to factor, and the dataframe
selection if multiple dataframe are present in one file (xlsx, rda).
}
\examples{
\dontrun{
 dataImport_demo()
}
}
\keyword{dataframe}
