% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportPlot.R
\name{exportPlot_demo}
\alias{exportPlot_demo}
\title{Export plot demo app}
\usage{
exportPlot_demo()
}
\value{
A Shiny APP.
}
\description{
R Shiny app demo to export plot
}
\details{
This demo allow to export an histogram plot from a
dataframe selected before hand.
}
\examples{
\donttest{
    exportPlot_demo()
}
}
\keyword{hplot}
