% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportPlot.R
\name{exportPlot_demo_2}
\alias{exportPlot_demo_2}
\title{Export plot server module}
\usage{
exportPlot_demo_2()
}
\value{
A Shiny app.
}
\description{
R Shiny module serve to export plot
}
\details{
This demo allow to export an heatmap plot from a
the example data MS_2.rda
}
\examples{
\dontrun{
    exportPlot_demo_2()
}
}
\keyword{hplot}
