% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectVar.R
\name{find_vars}
\alias{find_vars}
\title{Function to select a variable fitting a criterion in a dataframe}
\usage{
find_vars(df, filter)
}
\arguments{
\item{df}{A dataframe.}

\item{filter}{A function to use as a filter (e.g is.numeric).}
}
\value{
the column fitting the filter.
}
\description{
Function to select a variable fitting a criterion in a dataframe
}
