% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImport.R
\name{get_dataframe}
\alias{get_dataframe}
\title{Get dataframe name}
\usage{
get_dataframe(file)
}
\arguments{
\item{file}{The file path}
}
\value{
A vector of all the dataframe name present.
}
\description{
Extract the name of the different dataframe present in a file
}
\details{
This function detect the extension of the file and extract if
necessary the different dataframe / sheet names available.
}
\keyword{dataframe}
