% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHist.R
\name{ggplot_truehist}
\alias{ggplot_truehist}
\title{Plot ggplot histogram}
\usage{
ggplot_truehist(data, breaks = 50, title = NULL)
}
\arguments{
\item{data}{A numeric vector of data to use to generate a histogram.}

\item{breaks}{Numeric. Number of bins. Default is 50}

\item{title}{Character. Plot title.}
}
\value{
A ggplot histogram with a title
}
\description{
Plot ggplot histogram
}
\author{
Max Qiu, Louis Le Nézet, Alyssa Obermayer, Jared Andrews
}
\keyword{histogram}
