% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{ggplotly_2Dpca}
\alias{ggplotly_2Dpca}
\title{plot PCA 2D}
\usage{
ggplotly_2Dpca(
  df,
  sample_anno,
  sample_anno_col,
  PC1,
  PC2,
  pal,
  title,
  font_size
)
}
\arguments{
\item{sample_anno}{A data frame containing sample metadata}

\item{sample_anno_col}{Character. Sample annotation column in `sample_anno`}

\item{title}{Character. Plot title}

\item{data}{A data frame with feature (row) by sample (column).}
}
\value{
GGplot PCA
}
\description{
plot PCA 2D
}
