% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{ggplotly_3Dpca}
\alias{ggplotly_3Dpca}
\title{Plot PCA 3D}
\usage{
ggplotly_3Dpca(df, sample_anno, sample_anno_col, pal, title)
}
\arguments{
\item{sample_anno}{A data frame containing sample metadata}

\item{sample_anno_col}{Character. Sample annotation column in `sample_anno`}

\item{title}{Character. Plot title}

\item{data}{A data frame with feature (row) by sample (column).}
}
\value{
Interactive PCA plot
}
\description{
Plot PCA 3D
}
