% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmaply.R
\name{plotHeatmaply}
\alias{plotHeatmaply}
\title{plot Heatmap}
\usage{
plotHeatmaply(
  df,
  type = c("heatmap", "correlation", "distance"),
  direction = ifelse(type == "heatmap", NULL, c("col", "row")),
  select_dist = ifelse(type == "distance", c("euclidean", "maximum", "manhattan",
    "canberra", "binary", "minkowski"), NULL),
  select_corr = ifelse(type == "correlation", c("pearson", "kendall", "spearman"), NULL),
  select_hclust = c("ward.D", "ward.D2", "single", "complete", "average", "mcquitty",
    "median", "centroid"),
  select_scale = ifelse(type == "heatmap", c("column", "row"), "none"),
  draw_dendrogram = c("none", "row", "column", "both"),
  showRowDendrogram = TRUE,
  showColDendrogram = TRUE,
  showRowLabel = TRUE,
  showColLabel = TRUE,
  fontsize_col,
  fontsize_row,
  pal,
  title
)
}
\arguments{
\item{df}{A data frame}

\item{type}{Character, choose from one of three values for plot types}

\item{select_dist}{Character}

\item{select_corr}{Character}

\item{select_scale}{Character}

\item{draw_dendrogram}{Character}

\item{showRowDendrogram}{Boolean}

\item{showColDendrogram}{Boolean}

\item{showRowLabel}{Boolean}

\item{showColLabel}{Boolean}

\item{pal}{Character}

\item{title}{Character}
}
\value{
A heatmap plot generated with pheatmap
}
\description{
plot Heatmap
}
