% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHist.R
\name{plotHist_demo}
\alias{plotHist_demo}
\title{Histogram plot demo app}
\usage{
plotHist_demo()
}
\value{
A shiny app
}
\description{
R Shiny module server to generate an histogram plot.
}
\details{
This module create a histogram plot app.
The user can select the dataframe and the variable to use,
as well as the number of bins for the histogramm.
}
\examples{
\dontrun{
    plotHist_demo()
}
}
\keyword{hplot}
