% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHist.R
\name{plotHist_ui}
\alias{plotHist_ui}
\title{Histogram plot ui module}
\usage{
plotHist_ui(id)
}
\arguments{
\item{id}{A string.}
}
\value{
A Shiny UI.
}
\description{
R Shiny module UI to generate an hitogram plot.
}
\details{
This module create a histogram plot from a given
dataframe.
The number of bins is selectable by the user.
}
\examples{
\dontrun{
    plotHist_demo()
}
}
\keyword{hplot}
