% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectVar.R
\name{selectVar_demo}
\alias{selectVar_demo}
\title{R Shiny demo app to select a variable that fits a criterion}
\usage{
selectVar_demo(filter = is.numeric)
}
\arguments{
\item{filter}{A function to use as a filter (default is.numeric).}
}
\value{
A Shiny app.
}
\description{
R Shiny demo app to select a variable that fits a criterion
}
\examples{
\dontrun{
selectVar_demo(filter = is.numeric)
}
}
