% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectVar.R
\name{selectVar_server}
\alias{selectVar_server}
\title{R Shiny module to select a variable that fits a criterion}
\usage{
selectVar_server(id, df, filter = is.numeric)
}
\arguments{
\item{id}{A string.}

\item{df}{A reactive dataframe.}

\item{filter}{A function to use as a filter (default is.numeric).}
}
\value{
A Shiny server.
}
\description{
R Shiny module to select a variable that fits a criterion
}
\details{
The choice are updated in the server with the find_vars
function.
}
\examples{
\donttest{
selectVar_demo()
}
}
