% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectVar.R
\name{selectVar_ui}
\alias{selectVar_ui}
\title{R Shiny module to select a variable that fits a criterion}
\usage{
selectVar_ui(id, label = "Choose a column")
}
\arguments{
\item{id}{A string.}

\item{label}{Title to prompt to the user.}
}
\value{
A Shiny UI.
}
\description{
R Shiny module to select a variable that fits a criterion
}
\examples{
\donttest{
selectVar_demo()
}
}
