% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{ggplotly_2Dpca}
\alias{ggplotly_2Dpca}
\title{R Shiny module to generate PCA plot with choice of 2D or 3D}
\usage{
ggplotly_2Dpca(
  df,
  sample_anno,
  sample_anno_col,
  PC1,
  PC2,
  pal,
  title,
  font_size
)
}
\arguments{
\item{df}{A dataframe containing the value for each sample (column)
for all the features (rows).}

\item{sample_anno}{A data frame containing sample metadata}

\item{sample_anno_col}{Character. Sample annotation column in `sample_anno`}

\item{title}{Character. Plot title}

\item{id}{A string.}

\item{data}{A data frame with feature (row) by sample (column).}
}
\value{
A Shiny module.

GGplot PCA
}
\description{
R Shiny module to generate PCA plot with choice of 2D or 3D
}
\examples{
plotPCA_demo()
plotPCA_demo_2()
plot PCA 2D

}
