% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmaply.R
\name{plotHeatmaply}
\alias{plotHeatmaply}
\title{R Shiny module to generate a heatmap/distance/correlation using `heatmaply`}
\usage{
plotHeatmaply(
  df,
  type,
  direction = NULL,
  select_dist = NULL,
  select_corr = NULL,
  select_scale = "none",
  pal,
  ...
)
}
\arguments{
\item{df}{A data frame}

\item{type}{Character to choose from one of three types: "heatmap",
"correlation", "distance".}

\item{direction}{Character taking the following values: "col" or "row" for
correlation or distance types plot.}

\item{select_dist}{Method used to compute the distance (dissimilarity)
between both rows and columns that can take the following values :
"euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski". To be
use with type distance.}

\item{select_corr}{Correlation analysis model to use that can take the
following values : "pearson", "kendall", "spearman" to be use with type
correlation.}

\item{select_scale}{Character indicating if the values should be centered and
scaled in either the row direction or the column direction, or none.
The default is "none".}

\item{pal}{Color palette to use. Obtained from
`RColorBrewer::brewer.pal.info`}

\item{...}{Other parameters to be passed to  `heatmaply()`}

\item{id}{A string.}
}
\value{
A Shiny module.

A heatmap plot generated with pheatmap
}
\description{
R Shiny module to generate a heatmap/distance/correlation using `heatmaply`
}
\examples{
plotHeatmaply_demo()
plotHeatmaply_demo_2()
plot Heatmap

}
