% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHist.R
\name{plotHist_ui}
\alias{plotHist_ui}
\title{Histogram plot ui module}
\usage{
plotHist_ui(id)
}
\arguments{
\item{id}{A string.}
}
\value{
A Shiny UI.
}
\description{
R Shiny module UI to generate a histogram plot.
}
\details{
This module create a histogram plot from a given
dataframe.
The number of bins is selectable by the user.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ui <- shiny::fluidPage(
      plotHist_ui("hist")
)

server <- function(input, output, session) {
  speed <- reactive(cars$speed)
  plotHist_server("hist", speed)
}
shiny::shinyApp(ui, server)
\dontshow{\}) # examplesIf}
}
\author{
Max Qiu, Louis Le Nézet, Alyssa Obermayer, Jared Andrews
}
\keyword{histogram}
